userdetail usrphoto addedd

+++ db/migrate/20181108220748_create_teams.rb
@@ -4,7 +4,7 @@ class CreateTeams < ActiveRecord::Migration[5.2]
       t.string :name
       t.string :department
       t.references :company, foreign_key: true
-      t.integer :owner_id
+      t.bigint :owner_id
 
       t.timestamps
     end
